Port to 1.18, use custom portal block on the client instead of tinting vanilla portals
- Vanilla portals are now unaffected by interdimensional tinting
- Textures (and by extension tinting and particles) don't work because the polymer block on the client uses the virtual block state
- Will be able to change this behaviour in a future version of polymer. If needed, another block can be made and registered client side only instead

+++ src/main/kotlin/net/quiltservertools/interdimensional/command/PortalCommand.kt
@@ -96,9 +96,9 @@ object PortalCommand : Command {
                         blockState.blockState.block,
                         destination,
                         sourceWorld,
-                        (rgb[0] * 255).toInt().toByte(),
-                        (rgb[1] * 255).toInt().toByte(),
-                        (rgb[2] * 255).toInt().toByte(),
+                        (rgb[0] * 255).toInt().toUByte(),
+                        (rgb[1] * 255).toInt().toUByte(),
+                        (rgb[2] * 255).toInt().toUByte(),
                         flat,
                         ignitionSource
                     )
@@ -110,9 +110,9 @@ object PortalCommand : Command {
                     blockState.blockState.block,
                     destination,
                     sourceWorld,
-                    0,
-                    0,
-                    0,
+                    (0).toUByte(),
+                    (0).toUByte(),
+                    (0).toUByte(),
                     flat,
                     ignitionSource
                 )
